added static cast to update method

+++ include/md5.hpp
@@ -159,7 +159,7 @@ namespace libcrypt{
         template<typename T>
         void update(const T& _data){
             for(const auto& i : _data){
-                data[datalen] = i;
+                data[datalen] = static_cast<std::uint8_t>(i);
                 datalen++;
                 if(datalen == 64){
                     transform();
